examples/neo4j_requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.1
langgraph>=0.0.36
neo4j>=5.14.0
mcp>=0.0.1
python-dotenv>=1.0.0
langgraph-supervisor>=0.1.0 